added check to make sure a OrderTxn is a work order before deleting any tasks associated to it removed inventory entries reln to work_effort

+++ erp_work_effort/app/models/extensions/order_txn.rb
@@ -3,8 +3,11 @@ OrderTxn.class_eval do
   before_destroy :destroy_work_items
 
   def destroy_work_items
-    root_txn.work_efforts.each do |work_effort|
-      work_effort.destroy
+  	# if this is a work order then we should destroy any related tasks
+    if BizTxnType.iid('work_order') && root_txn.biz_txn_type.is_descendant_of?(BizTxnType.iid('work_order'))
+      root_txn.work_efforts.each do |work_effort|
+        work_effort.destroy
+      end
     end
   end
